pyproject.toml
[project]
name = "openclaw-voice-bridge"
version = "2.0.0"
description = "Pipecat-based voice interface for OpenClaw"
requires-python = ">=3.11"
dependencies = [
    "pipecat-ai[webrtc,groq,silero,elevenlabs]>=0.0.68",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
